requirements.txt
drfscaffolding==0.2.1
psycopg2
djangorestframework-xml
